+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.4
+
+- More convenient drawing (one line of code to set up)
+- Offscreen rendering helper
+
 ## 0.0.1
 
 - First version; most core functionality